Project restructuring
Based on https://github.com/golang-standards/project-layout

Main application in /cmd/ginvalid
  - builds binary 'ginvalid' by default

Internal library code in /internal

Import paths adjusted accordingly.
