+++ cmd/ginvalid/main.go
@@ -10,10 +10,10 @@ import (
 	"os/signal"
 	"strings"
 
-	"github.com/G-Node/gin-valid/config"
-	"github.com/G-Node/gin-valid/helpers"
-	"github.com/G-Node/gin-valid/log"
-	"github.com/G-Node/gin-valid/web"
+	"github.com/G-Node/gin-valid/internal/config"
+	"github.com/G-Node/gin-valid/internal/helpers"
+	"github.com/G-Node/gin-valid/internal/log"
+	"github.com/G-Node/gin-valid/internal/web"
 	"github.com/docopt/docopt-go"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
