+++ internal/helpers/handleShell.go
