+++ internal/resources/svg.go
