+++ internal/resources/templates/bids_results.go
