+++ internal/resources/templates/fail.go
