+++ internal/resources/templates/layout.go
