+++ internal/resources/templates/pubvalidate.go
