+++ internal/resources/templates/repolist.go
