+++ internal/resources/templates/root.go
