+++ internal/web/results.go
@@ -11,10 +11,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/G-Node/gin-valid/config"
-	"github.com/G-Node/gin-valid/helpers"
-	"github.com/G-Node/gin-valid/log"
-	"github.com/G-Node/gin-valid/resources/templates"
+	"github.com/G-Node/gin-valid/internal/config"
+	"github.com/G-Node/gin-valid/internal/helpers"
+	"github.com/G-Node/gin-valid/internal/log"
+	"github.com/G-Node/gin-valid/internal/resources/templates"
 	"github.com/gorilla/mux"
 )
