+++ internal/web/user.go
@@ -20,10 +20,10 @@ import (
 	gcfg "github.com/G-Node/gin-cli/ginclient/config"
 	glog "github.com/G-Node/gin-cli/ginclient/log"
 	gweb "github.com/G-Node/gin-cli/web"
-	"github.com/G-Node/gin-valid/config"
-	"github.com/G-Node/gin-valid/helpers"
-	"github.com/G-Node/gin-valid/log"
-	"github.com/G-Node/gin-valid/resources/templates"
+	"github.com/G-Node/gin-valid/internal/config"
+	"github.com/G-Node/gin-valid/internal/helpers"
+	"github.com/G-Node/gin-valid/internal/log"
+	"github.com/G-Node/gin-valid/internal/resources/templates"
 	gogs "github.com/gogits/go-gogs-client"
 	"github.com/gorilla/mux"
 )
